refactor: make code styled like more javascript more than csharp.
