+++ noname/init/onload/index.js
@@ -1,0 +1,4 @@
+import { DefaultSplash } from "./default-splash.js";
+import { WideSplash } from "./wide-splash.js";
+
+export const defaultSplashs = [new DefaultSplash(), new WideSplash()];
