--- noname/init/onload/mod.js
@@ -1,4 +1,0 @@
-import { DefaultSplash } from "./DefaultSplash.js";
-import { WideSplash } from "./WideSplash.js";
-
-export const defaultSplashs = [new DefaultSplash(), new WideSplash()];
