+++ noname/init/onload/wide-splash.d.ts
@@ -1,0 +1,8 @@
+import { OnloadSplash } from "./onload-splash";
+
+export class WideSplash extends DefaultSplash implements OnloadSplash {
+	readonly id: string;
+	readonly name: string;
+
+	protected path: string;
+}
